+++ mopidy_mpd/__init__.py
@@ -1,4 +1,3 @@
-import logging
 import pathlib
 
 import pkg_resources
@@ -6,9 +5,6 @@ import pkg_resources
 from mopidy import config, ext
 
 __version__ = pkg_resources.get_distribution("Mopidy-MPD").version
-
-# TODO: If you need to log, use loggers named after the current Python module
-logger = logging.getLogger(__name__)
 
 
 class Extension(ext.Extension):
@@ -22,25 +18,17 @@ class Extension(ext.Extension):
 
     def get_config_schema(self):
         schema = super().get_config_schema()
-        # TODO: Comment in and edit, or remove entirely
-        #schema["username"] = config.String()
-        #schema["password"] = config.Secret()
+        schema["hostname"] = config.Hostname()
+        schema["port"] = config.Port(optional=True)
+        schema["password"] = config.Secret(optional=True)
+        schema["max_connections"] = config.Integer(minimum=1)
+        schema["connection_timeout"] = config.Integer(minimum=1)
+        schema["zeroconf"] = config.String(optional=True)
+        schema["command_blacklist"] = config.List(optional=True)
+        schema["default_playlist_scheme"] = config.String()
         return schema
 
     def setup(self, registry):
-        # You will typically only implement one of the following things
-        # in a single extension.
+        from .actor import MpdFrontend
 
-        # TODO: Edit or remove entirely
-        from .frontend import FoobarFrontend
-        registry.add("frontend", FoobarFrontend)
-
-        # TODO: Edit or remove entirely
-        from .backend import FoobarBackend
-        registry.add("backend", FoobarBackend)
-
-        # TODO: Edit or remove entirely
-        registry.add("http:static", {
-            "name": self.ext_name,
-            "path": str(pathlib.Path(__file__) / "static")),
-        })
+        registry.add("frontend", MpdFrontend)