+++ CHANGELOG.rst
@@ -4,6 +4,13 @@ Changelog
 
 
 v3.0.0rc1 (UNRELEASED)
-========================================
+======================
 
-- Initial release.
+Initial release which extracts the Mopidy-MPD extension from Mopidy core.
+
+This is an alpha release because it depends on the pre-releases of Mopidy 3.0.
+
+Changes since Mopidy 2.3:
+
+- Improved the ``swap`` command to swap the tracks without rebuilding
+  the full tracklist.